requirements.txt
PyGithub
matplotlib
pandas
python-dateutil
requests
